create module card
