+++ Todo-list/js/modules/main.js
@@ -1,5 +1,7 @@
 import {getLocalData, setLocalData, removeLocalData} from './userData.js'
-import {cssWrapper, cssPanel, cssPanelButton, cssPanelInput, cssCard, cssCardStatusWraper, cssStatusSpan, cssCardStatus, cssCardDate, cssCardText, cssCardCloseButton, cssImage, setCardStyle, setHover, setFocus} from './style.js'
+import {cssWrapper, cssPanel, cssPanelButton, cssPanelInput, cssRoot} from './style.js'
+import {setHover, setFocus} from './events.js';
+import {renderCard} from './card.js';
 
 export let main = function (){
 
@@ -12,30 +14,9 @@ export let main = function (){
         main = getLocalData();
     }
 
-
     let completed = main.filter(item => item.status);
     let searched = [];
 
-
-//functions
-const deleteCard = function(){
-    let result = confirm("Delete Card?");
-    if (result){
-        let parent = this.closest(".card")
-            let newMain = main.filter(item => item.id != parent.getAttribute('data-id'));
-            main = [...newMain];
-            let newCompleted = completed.filter(item => item.id != parent.getAttribute('data-id'));
-            completed = [...newCompleted].sort((a,b) => b.id - a.id);
-
-        parent.remove();
-        if(main.length == 0){
-            removeLocalData();
-        }
-        setCounterAllCards();
-        setCounterCompletedCards();
-        setLocalData(main);
-    }
-}
 
 const deleteAllCards = function(){
     let result = confirm("Delete All Cards?");
@@ -47,8 +28,7 @@ const deleteAllCards = function(){
         setCounterCompletedCards()
         removeLocalData();
     }
-
-;}
+};
 
 const deleteLastCard = function(){
     let result = confirm("Delete Last Card?");
@@ -75,14 +55,14 @@ const deleteLastCard = function(){
         setCounterAllCards();
         setCounterCompletedCards();
     }
-}
+};
 
 
 
-let setCounterAllCards = function(){
+const setCounterAllCards = function(){
     counterAllCards.innerHTML = `All: <span>${main.length}</span>`;
 }
-let setCounterCompletedCards = function(){
+const setCounterCompletedCards = function(){
     counterCompletedCards.innerHTML = `Completed: <span>${completed.length}</span>`; 
 
 }
@@ -124,93 +104,29 @@ const addCard = function(){
     } 
 }
 
-const createCard = function (item) {
-    let card = document.createElement("div");
-        card.classList.add("card");
-        card.style.cssText = cssCard;
-        card.setAttribute("data-id", item.id);
+const deleteCard = function(){
+    let result = confirm("Delete Card?");
+    if (result){
+        let parent = this.closest(".card")
+            let newMain = main.filter(item => item.id != parent.getAttribute('data-id'));
+            main = [...newMain];
+            let newCompleted = completed.filter(item => item.id != parent.getAttribute('data-id'));
+            completed = [...newCompleted].sort((a,b) => b.id - a.id);
 
-    let cardStatusWraper = document.createElement("label");
-        cardStatusWraper.style.cssText = cssCardStatusWraper;
-
-    let cardStatusSpan = document.createElement("span");
-        cardStatusSpan.style.cssText = cssStatusSpan;
-
-    let cardStatus = document.createElement("input");
-        cardStatus.setAttribute("type", "checkbox");
-        cardStatus.checked = item.status;
-        cardStatus.classList.add("card--status");
-        cardStatus.style.cssText = cssCardStatus;
-        if (item.status){
-            card.style.backgroundColor= "gray";
-            cardStatusSpan.style.backgroundImage = cssImage;
-
-            let key =  completed.some( elem => elem.id === item.id);
-            if (key == false){
-                completed.push(item);
-            }
+        parent.remove();
+        if(main.length == 0){
+            removeLocalData();
         }
-        cardStatus.addEventListener("focus", function(){
-            cardStatusSpan.style.boxShadow = "0 0 0 0.15em DodgerBlue";
-        })
-        cardStatus.addEventListener("focusout", function(){
-            cardStatusSpan.style.boxShadow = "none"
-        })
-        cardStatus.addEventListener("focus", function(){
-            card.style.boxShadow = "0 0 0 0.04em black";
-        })
-        cardStatus.addEventListener("focusout", function(){
-            card.style.boxShadow = "none"
-        })
-
-
-
-    let cardDate = document.createElement("div");
-        cardDate.classList.add("card--date");
-        cardDate.textContent = item.date
-        cardDate.style.cssText = cssCardDate
-
-    let cardText = document.createElement("div");
-        cardText.classList.add("card--text");
-        cardText.textContent = item.task
-        cardText.style.cssText = cssCardText
-        if (item.status){
-            cardText.style.textDecoration = "line-through";
-        }
-
-    let cardCloseButton = document.createElement("div");
-        cardCloseButton.classList.add("card--date");
-        cardCloseButton.textContent = "X"
-        cardCloseButton.style.cssText = cssCardCloseButton;
-        cardCloseButton.addEventListener("click", deleteCard);
-        setHover("Red", "0.4", cardCloseButton);
-        
-
-        cardStatus.addEventListener("click",setCardStyle);
-        cardStatus.addEventListener("click", function(){
-             item.status = cardStatus.checked;
-             if(item.status){
-                completed.push(item);
-             }else{
-                let parent = this.closest(".card")
-                let newCompleted = completed.filter(item => item.id != parent.getAttribute('data-id'));
-                completed.length = 0;
-                completed = [...newCompleted].sort((a,b) => b.id - a.id);
-             }
-     
-            setCounterCompletedCards();
-            setLocalData(main);
-        });
-        setFocus(cardStatus)
-
-        cardsAera.prepend(card);
-        card.append(cardStatusWraper, cardText, cardCloseButton, cardDate); 
-        cardStatusWraper.append(cardStatusSpan, cardStatus);
+        setCounterAllCards();
+        setCounterCompletedCards();
+        setLocalData(main);
+    }
 }
+
 
 const createCardsAera = function (arr){ 
     arr.forEach((item) => {
-        createCard(item);
+        renderCard(item, main, completed, cardsAera, deleteCard, setCounterCompletedCards,setLocalData);
     })
     setCounterAllCards();
     setCounterCompletedCards(); 
@@ -218,14 +134,12 @@ const createCardsAera = function (arr){
 
 
 let root = document.getElementById("root")
-    root.style.cssText = `
-            padding: 10px;
-`;
-
+    root.style.cssText = cssRoot;
+ 
 let wrapper = document.createElement("div");
     wrapper.classList.add("wrapper");
     wrapper.style.cssText =  cssWrapper;
-
+ 
 let panelActions = document.createElement("div");
     panelActions.classList.add("panel-card-actions");
     panelActions.style.cssText = cssPanel;
@@ -318,4 +232,4 @@ root.append(wrapper);
 wrapper.append(panelActions, panelCardInfo, cardsAera);
 panelActions.append(buttonDeleteAllCards, buttonDeleteLastCard, inputTask, buttonAddCard);
 panelCardInfo.append(counterAllCards, counterCompletedCards, buttonShowAll,buttonShowCompleted, inputSearch);
-}
+ };
