+++ Todo-list/js/modules/userData.js
@@ -11,25 +11,4 @@ export let setLocalData = function(arr){
 export let removeLocalData = function(){
     localStorage.removeItem("todo");
 }
-/*
-[
-    {
-        "id": 1,
-        "task": "первый",
-        "status": false,
-        "date": "13 - 11 - 2022 "
-    },
-    {
-        "id": 2,
-        "task": "второй",
-        "status": false,
-        "date": "13 - 11 - 2022 "
-    },
-    {
-        "id": 3,
-        "task": "третий",
-        "status": false,
-        "date": "13 - 11 - 2022 "
-    }
-]
-*/
+
